+++ src/physics/matter-js/components/SetBody.js
@@ -125,7 +125,10 @@ var SetBody = {
             this.world.add(this.body);
         }
 
-        this.setOrigin(body.render.sprite.xOffset, body.render.sprite.yOffset);
+        if (this._originComponent)
+        {
+            this.setOrigin(body.render.sprite.xOffset, body.render.sprite.yOffset);
+        }
 
         return this;
     },